+++ com-services-api/src/main/java/com/services/api/form/orderdetail/CreateOrderDetailForm.java
@@ -20,5 +20,5 @@ public class CreateOrderDetailForm {
     private Integer orderDetailQuantity;
     @NotNull(message = "orderDetailPrice cannot be null")
     @ApiModelProperty(name = "orderDetailPrice")
-    private Long orderDetailPrice; 
+    private Integer orderDetailPrice; 
 }
